Add Job Info command and moved copyJobId command to connection commands
