+++ src/commands/actions.ts
@@ -1,5 +1,5 @@
 import path from "path";
-import { commands, Disposable, env, l10n, TreeItem, Uri, window, WorkspaceFolder } from "vscode";
+import { commands, Disposable, l10n, TreeItem, Uri, window, WorkspaceFolder } from "vscode";
 import IBMi from "../api/IBMi";
 import { Tools } from "../api/Tools";
 import Instance from "../Instance";
@@ -127,7 +127,7 @@ export function registerActionsCommands(instance: Instance): Disposable[] {
         ext: undefined
       };
 
-      let inputPath: string | undefined      
+      let inputPath: string | undefined
       const connection = instance.getConnection()!; //safe as action requires to be connected
 
       if (options.qualifiedObject) {
@@ -137,7 +137,7 @@ export function registerActionsCommands(instance: Instance): Disposable[] {
         // Value collected from user input
 
         let initialPath = ``;
-        const editor = window.activeTextEditor;        
+        const editor = window.activeTextEditor;
 
         if (editor && connection) {
           const config = connection.getConfig();
@@ -181,17 +181,6 @@ export function registerActionsCommands(instance: Instance): Disposable[] {
           const nameDetail = path.parse(object);
           refreshDiagnosticsFromServer(connection, [{ library, object: nameDetail.name, extension: (nameDetail.ext.length > 1 ? nameDetail.ext.substring(1) : undefined), workspace: options.workspace }], options.keepDiagnostics);
         }
-      }
-    }),
-
-    commands.registerCommand(`code-for-ibmi.copyJobId`, async () => {
-      const connection = instance.getConnection();
-      const sqlJobId = connection?.getSqlJobId();
-      if (sqlJobId) {
-        await env.clipboard.writeText(sqlJobId);
-        window.showInformationMessage(`Job ID copied: ${sqlJobId}`);
-      } else {
-        window.showWarningMessage(`No job ID available`);
       }
     }),
   ]
